requirements.txt
streamlit>=1.31.0
opencv-python-headless>=4.8.0
numpy>=1.26.0
Pillow>=10.0.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
pandas>=1.3.0
deepface>=0.0.79
tensorflow==2.15.0
keras==2.15.0
scipy>=1.7.0
mtcnn>=0.1.0
retina-face>=0.0.1
requests>=2.25.0
dlib-binary>=19.22.0